# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PharmaPulse API",
    "description": "A pharmacy's sales history becomes a purchase quantity, with the odds and the cost of being wrong attached. See CONTRACTS.md section C3.",
    "version": "1.0.0"
  },
  "paths": {
    "/api/health": {
      "get": {
        "tags": [
          "ops"
        ],
        "summary": "Health",
        "operationId": "health_api_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Health Api Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/metrics": {
      "get": {
        "tags": [
          "ops"
        ],
        "summary": "Metrics",
        "operationId": "metrics_api_metrics_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Metrics Api Metrics Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/series": {
      "get": {
        "tags": [
          "forecast"
        ],
        "summary": "Series",
        "description": "Catalogue with the measured demand class, which drives a UI chip.",
        "operationId": "series_api_series_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Series Api Series Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/history": {
      "get": {
        "tags": [
          "forecast"
        ],
        "summary": "History",
        "operationId": "history_api_history_get",
        "parameters": [
          {
            "name": "series_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "N02BE",
              "title": "Series Id"
            }
          },
          {
            "name": "grain",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "week",
              "title": "Grain"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2500,
              "minimum": 1,
              "default": 120,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response History Api History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/forecast": {
      "get": {
        "tags": [
          "forecast"
        ],
        "summary": "Forecast",
        "operationId": "forecast_api_forecast_get",
        "parameters": [
          {
            "name": "series_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "N02BE",
              "title": "Series Id"
            }
          },
          {
            "name": "grain",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "week",
              "title": "Grain"
            }
          },
          {
            "name": "horizon",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 8,
              "title": "Horizon"
            }
          },
          {
            "name": "with_members",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "With Members"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Forecast Api Forecast Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/explain": {
      "get": {
        "tags": [
          "forecast"
        ],
        "summary": "Explain",
        "operationId": "explain_api_explain_get",
        "parameters": [
          {
            "name": "series_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "R06",
              "title": "Series Id"
            }
          },
          {
            "name": "grain",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "month",
              "title": "Grain"
            }
          },
          {
            "name": "horizon",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 6,
              "minimum": 1,
              "default": 1,
              "title": "Horizon"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Explain Api Explain Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/recommend": {
      "post": {
        "tags": [
          "decision"
        ],
        "summary": "Recommend",
        "operationId": "recommend_api_recommend_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Recommend Api Recommend Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/risk": {
      "get": {
        "tags": [
          "decision"
        ],
        "summary": "Risk",
        "operationId": "risk_api_risk_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Risk Api Risk Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/positions": {
      "get": {
        "tags": [
          "decision"
        ],
        "summary": "Positions",
        "description": "Live shelf position per product - what the dashboard opens on.",
        "operationId": "positions_api_positions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Positions Api Positions Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings": {
      "get": {
        "tags": [
          "decision"
        ],
        "summary": "Get Settings",
        "operationId": "get_settings_api_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Settings Api Settings Get"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "decision"
        ],
        "summary": "Put Settings",
        "description": "Lane 2. Editable, and never seen by the trainer.",
        "operationId": "put_settings_api_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "title": "Patch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Put Settings Api Settings Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/orders": {
      "post": {
        "tags": [
          "decision"
        ],
        "summary": "Commit Order",
        "description": "Accept or override a recommendation. Overrides need a reason.\n\nAppends to a hash-chained log: each entry stores the previous entry's hash,\nso an edit or deletion is detectable. The system recommends; a person commits.",
        "operationId": "commit_order_api_orders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCommit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Commit Order Api Orders Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/replay/start": {
      "post": {
        "tags": [
          "replay"
        ],
        "summary": "Start",
        "operationId": "start_api_replay_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Start Api Replay Start Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/replay/tick": {
      "post": {
        "tags": [
          "replay"
        ],
        "summary": "Tick",
        "operationId": "tick_api_replay_tick_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TickRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Tick Api Replay Tick Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/replay/stop": {
      "post": {
        "tags": [
          "replay"
        ],
        "summary": "Stop",
        "operationId": "stop_api_replay_stop_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TickRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Stop Api Replay Stop Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/replay/business-case": {
      "get": {
        "tags": [
          "replay"
        ],
        "summary": "Business Case",
        "description": "Our policy against a min/max policy, over the identical real days.\n\nSame data, same costs, same lead time, same review cadence, same protection\ninterval. The ONLY difference is that min/max sizes against the mean while\nwe size against the quantile the pharmacy's own cost ratio implies.\n\nThat makes the difference attributable to the thesis rather than to a\nhandicap given to the baseline.",
        "operationId": "business_case_api_replay_business_case_get",
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "2019-01-01",
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "2019-03-31",
              "title": "End Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Business Case Api Replay Business Case Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/": {
      "get": {
        "summary": "Root",
        "operationId": "root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Root  Get"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "OrderCommit": {
        "properties": {
          "series_id": {
            "type": "string",
            "title": "Series Id"
          },
          "recommended": {
            "type": "integer",
            "title": "Recommended"
          },
          "accepted": {
            "type": "integer",
            "title": "Accepted"
          },
          "service_level": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Level"
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": ""
          },
          "actor": {
            "type": "string",
            "title": "Actor",
            "default": "demo"
          }
        },
        "type": "object",
        "required": [
          "series_id",
          "recommended",
          "accepted"
        ],
        "title": "OrderCommit"
      },
      "RecommendRequest": {
        "properties": {
          "series_id": {
            "type": "string",
            "title": "Series Id",
            "default": "N02BE"
          },
          "service_level": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 0.999,
                "minimum": 0.01
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Level"
          },
          "lead_time_days": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 90.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Lead Time Days"
          },
          "stock_on_hand": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Stock On Hand"
          },
          "pack_size": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Pack Size"
          },
          "unit_cost": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Cost"
          },
          "unit_margin": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Margin"
          },
          "holding_cost_rate": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 5.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Holding Cost Rate"
          },
          "expiry_risk_rate": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiry Risk Rate"
          },
          "review_period_days": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 90.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Period Days"
          }
        },
        "type": "object",
        "title": "RecommendRequest"
      },
      "StartRequest": {
        "properties": {
          "from": {
            "type": "string",
            "title": "From",
            "default": "2019-01-01"
          },
          "to": {
            "type": "string",
            "title": "To",
            "default": "2019-03-31"
          },
          "policy": {
            "type": "string",
            "title": "Policy",
            "default": "pharmapulse"
          }
        },
        "type": "object",
        "title": "StartRequest"
      },
      "TickRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "steps": {
            "type": "integer",
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Steps",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "session_id"
        ],
        "title": "TickRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}